Framework for STS lookups.
Use wget to download each domain's policy file.

Cache domain policy files in $localstatedir/sts.

Provide a library to retrieve the cached policy file, if available, or to
download and cache it, safely. Provide the means to purge cached files,
when there are too many of them.

Testmxlookup reports the domain's policy, in addition to its MXes.

Add options to testmxlookup to show the domain's policy file.